Load and lay out most of the desktop layout
The challenge starts with a basic index.html and additional images that
can be loaded and laid out. There are files in the design directory
and a style-guide.md that gives a hint on what is needed to complete
the challenge. The direction of completion is up to the individual.

The index.html uses semantic tags as part of the challenge
instructions. The style.css is the file that controls style and layout
of the pages. The fonts and icons are external files that had to be
included.  The fonts are easier, at this point, to find and include
then the icons.