Add mobile layout and styles directory
The desktop layout was moved to a media query to switch to a mobile
first approach. Afterwards, a mobile layout was formed and partly
organized. While the mobile layout was being laid out, styles.css was
renamed style-mobile.css and normalized.css was added. Both are then
moved to a styles directory and index.html was updated to point to
them.